Add extra models
Add the model from Wang and Rosowsky (2012) and a simple model of pressure
deficit and latitude. The Powell et al. model remains the best model,
based on R-squared and AIC/BIC statistics. Conclusions are based on a GLS
regression, so implementing into a modelling framework will require
methods to handle the case where the error covariance matrix is not
positive definite.